// src/lib/types.d.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Asset, Entry } from 'contentful';
import { Document } from '@contentful/rich-text-types';

export interface ICategoriaFields {
	/** Nome categoria */
	nomeCategoria: string;

	/** Slug */
	slug: string;

	/** immagine */
	immagine?: Asset | undefined;

	/** Ordine */
	ordine: number;
}

export interface ICategoria extends Entry<ICategoriaFields> {
	sys: {
		id: string;
		type: string;
		createdAt: string;
		updatedAt: string;
		locale: string;
		contentType: {
			sys: {
				id: 'categoria';
				linkType: 'ContentType';
				type: 'Link';
			};
		};
	};
}

export interface IProdottoFields {
	/** Nome */
	nome: string;

	/** Prezzo */
	prezzo?: number | undefined;

	/** Foto */
	foto?: Asset | undefined;

	/** Categoria */
	categoria?: Entry<{ [fieldId: string]: unknown }> | undefined;
}

export interface IProdotto extends Entry<IProdottoFields> {
	sys: {
		id: string;
		type: string;
		createdAt: string;
		updatedAt: string;
		locale: string;
		contentType: {
			sys: {
				id: 'prodotto';
				linkType: 'ContentType';
				type: 'Link';
			};
		};
	};
}

export type CONTENT_TYPE = 'categoria' | 'prodotto';

export type LOCALE_CODE = 'en' | 'it';

export type CONTENTFUL_DEFAULT_LOCALE_CODE = 'it';